requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.0.0
pytorch-ignite>=0.4.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
